by this commit, all the get request endpoints has been finished, next stage, post and put methods

+++ models/engine/db_storage.py
@@ -15,12 +15,13 @@ from models.teacher import Teacher
 from models.lesson import Lesson
 from models.state import State
 from models.city import City
+from models.student import Student
 
 
 
 classes = {'Institution': Institution, 'Subject': Subject,
            'Teacher': Teacher, 'Lesson': Lesson, 'Class': Clas,
-           'State': State, 'City': City}
+           'State': State, 'City': City, 'Student': Student}
 
 
 class DBStorage:
@@ -35,21 +36,21 @@ class DBStorage:
         ACME_MYSQL_HOST = getenv('h')
         ACME_MYSQL_DB = getenv('D')
         ACME_ENV = getenv('mode')
-        # if None in (ACME_MYSQL_USER, ACME_MYSQL_PWD,
-                    # ACME_MYSQL_HOST, ACME_MYSQL_DB):
-            # string = "One or more required environment variables are not set."
-            # raise ValueError(string)
-
-        self.__engine = create_engine('mysql+mysqldb://{}:{}@{}/{}'.
-                                      format('root',
-                                             '' ,
-                                             'localhost',
-                                             ACME_MYSQL_DB))
-        # self.__engine = create_engine('mysql+mysqldb://{}:{}@{}/{}'.
-                                      # format(ACME_MYSQL_USER,
-                                             # ACME_MYSQL_PWD,
-                                             # ACME_MYSQL_HOST,
+        # print(ACME_MYSQL_USER, ACME_MYSQL_PWD, ACME_MYSQL_HOST, ACME_MYSQL_DB)
+        if None in (ACME_MYSQL_USER, ACME_MYSQL_PWD,
+                    ACME_MYSQL_HOST, ACME_MYSQL_DB):
+            string = "One or more required environment variables are not set."
+            raise ValueError(string)
+#         self.__engine = create_engine('mysql+mysqldb://{}:{}@{}/{}'.
+                                      # format('root',
+                                             # '' ,
+                                             # 'localhost',
                                              # ACME_MYSQL_DB))
+        self.__engine = create_engine('mysql+mysqldb://{}:{}@{}/{}'.
+                                      format(ACME_MYSQL_USER,
+                                             ACME_MYSQL_PWD,
+                                             ACME_MYSQL_HOST,
+                                             ACME_MYSQL_DB))
         if ACME_ENV == "test":
             Base.metadata.drop_all(self.__engine)
 
@@ -119,4 +120,9 @@ class DBStorage:
         return count
 
     def query(self, cls):
+        """ give access to the current session outside this module"""
         return self.__session.query(cls)
+
+    def rollback(self):
+        """rollback"""
+        self.__session.rollback()
